+++ Client/src/App.jsx
@@ -1,14 +1,12 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route, Link } from "react-router-dom";
-import AdminViewUserProfile from "./pages/AdminProfileUser";
-import Sidebar from "./pages/Sidebar";   
-import UploadData from "./pages/UploadMemberData";
+
+import MainPageUpload from "./pages/UploadPage/MainPageUpload";
 import GroupDetailsPage from "./pages/GroupDetailsPage/GroupDetailsPage"
 import UserSearchComponent from "./pages/Searchpage/UserSearchComponent";
-import UserProfileForm from "./pages/UserProfileForm";
-import CommunityMembersPage from "./pages/CommunityPage";
+// import UserProfileForm from "./pages/UserProfileForm"; 
 import EventList from "./pages/EventPages/EventList";
-
+import MainPageEditableUserProfile from "./pages/UserProfile/MainPageEditableUserProfile";
 import "./pages/index.css"; 
 import CvTemporaryPage from "./pages/cv/CvTemporaryPage";
 
@@ -19,19 +17,19 @@ function App() {
       <Routes>
         <Route path="/cv-temp" element={<CvTemporaryPage />} />
 
-       <Route path="/member/:id/data/" element={<UserProfileForm />} />
-       <Route path="/member/data/" element={<UserProfileForm />} />
+       <Route path="/member/:id/data/" element={<MainPageEditableUserProfile />} />
+       <Route path="/member/data/" element={<MainPageEditableUserProfile />} />
 
         <Route path="/UserSearch" element={<UserSearchComponent />} />
-        <Route path="/manager/group/:idGroup" element={<CommunityMembersPage />} />
-        <Route path="/member" element={<UploadData />} />
-        <Route path="/sidebar" element={<Sidebar />} />
-        <Route path="/manager/members/:id" element={<AdminViewUserProfile />} />
+        {/* <Route path="/manager/group/:idGroup" element={<CommunityMembersPage />} /> */}
+        <Route path="/member" element={<MainPageUpload />} />
+        {/* <Route path="/sidebar" element={<Sidebar />} /> */}
+        {/* <Route path="/manager/members/:id" element={<AdminViewUserProfile />} /> */}
 
         <Route path="/group/:groupId" element={<GroupDetailsPage />} />
         <Route path="/events" element={<EventList />} />
 
-       {<Route path="/" element={<UserProfileForm />} />}
+       {<Route path="/" element={<MainPageEditableUserProfile />} />}
         {/* <Route path="/member/data" element={<UploadData />} /> */}
          {/* <Route path="/" element={<Slider />} /> */}
         {/* <Route path="/manager/member" element={<UploadProfile />} /> */}
