+++ Client/src/pages/Searchpage/SearchField.jsx
@@ -1,4 +1,3 @@
-// קומפוננט שדה חיפוש
 import React from 'react';
 
 import { Search } from 'lucide-react';
