+++ Client/src/pages/UploadPage/LinkedInSection.jsx
@@ -1,0 +1,44 @@
+import React from 'react';
+
+/**
+ * LinkedInSection Component
+ * Handles LinkedIn URL input
+ * @param {string} linkedinUrl - Current LinkedIn URL value
+ * @param {Function} setLinkedinUrl - Function to update LinkedIn URL
+ * @param {boolean} isLoading - Loading state
+ */
+function LinkedInSection({ linkedinUrl, setLinkedinUrl, isLoading }) {
+  return (
+    <div className="mb-8">
+      <div className="flex items-center mb-4">
+        <div className="w-12 h-12 bg-blue-100 rounded-full flex items-center justify-center mr-4">
+          <svg className="w-6 h-6 text-blue-600" fill="currentColor" viewBox="0 0 24 24">
+            <path d="M20.447 20.452h-3.554v-5.569c0-1.328-.027-3.037-1.852-3.037-1.853 0-2.136 1.445-2.136 2.939v5.667H9.351V9h3.414v1.561h.046c.477-.9 1.637-1.85 3.37-1.85 3.601 0 4.267 2.37 4.267 5.455v6.286zM5.337 7.433c-1.144 0-2.063-.926-2.063-2.065 0-1.138.92-2.063 2.063-2.063 1.14 0 2.064.925 2.064 2.063 0 1.139-.925 2.065-2.064 2.065zm1.782 13.019H3.555V9h3.564v11.452zM22.225 0H1.771C.792 0 0 .774 0 1.729v20.542C0 23.227.792 24 1.771 24h20.451C23.2 24 24 23.227 24 22.271V1.729C24 .774 23.2 0 22.222 0h.003z"/>
+          </svg>
+        </div>
+        <div>
+          <h3 className="text-xl font-semibold text-gray-800">Import from LinkedIn</h3>
+          <p className="text-gray-600">Automatically fill your profile with LinkedIn data</p>
+        </div>
+      </div>
+      
+      <div className="relative">
+        <input
+          type="url"
+          className="w-full px-4 py-4 border-2 border-gray-200 rounded-xl focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all text-lg"
+          placeholder="https://www.linkedin.com/in/your-profile"
+          value={linkedinUrl}
+          onChange={(e) => setLinkedinUrl(e.target.value)}
+          disabled={isLoading}
+        />
+        <div className="absolute inset-y-0 right-0 flex items-center pr-4">
+          <svg className="w-6 h-6 text-blue-500" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M13.828 10.172a4 4 0 00-5.656 0l-4 4a4 4 0 105.656 5.656l1.102-1.101m-.758-4.899a4 4 0 005.656 0l4-4a4 4 0 00-5.656-5.656l-1.1 1.1" />
+          </svg>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default LinkedInSection;
