+++ Client/src/pages/UserProfile/ContactTab.jsx
@@ -1,0 +1,65 @@
+import React from 'react';
+
+/**
+ * ContactTab Component
+ * Renders contact information form fields
+ * @param {Object} formData - User form data
+ * @param {Function} handleChange - Form field change handler
+ */
+function ContactTab({ formData, handleChange }) {
+  return (
+    <div className="space-y-6">
+      <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
+        <div>
+          <label className="block text-sm font-semibold text-gray-700 mb-2">Phone Number</label>
+          <input
+            type="tel"
+            name="phone"
+            className="w-full px-4 py-3 border border-gray-300 rounded-xl focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all"
+            placeholder="Your phone number"
+            value={formData.phone}
+            onChange={handleChange}
+          />
+        </div>
+        <div>
+          <label className="block text-sm font-semibold text-gray-700 mb-2">Email Address</label>
+          <input
+            type="email"
+            name="email"
+            className="w-full px-4 py-3 border border-gray-300 rounded-xl focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all"
+            placeholder="your.email@example.com"
+            value={formData.email}
+            onChange={handleChange}
+          />
+        </div>
+      </div>
+      
+      <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
+        <div>
+          <label className="block text-sm font-semibold text-gray-700 mb-2">LinkedIn Profile</label>
+          <input
+            type="url"
+            name="linkedin_url"
+            className="w-full px-4 py-3 border border-gray-300 rounded-xl focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all"
+            placeholder="https://linkedin.com/in/yourprofile"
+            value={formData.linkedin_url}
+            onChange={handleChange}
+          />
+        </div>
+        <div>
+          <label className="block text-sm font-semibold text-gray-700 mb-2">Facebook Profile</label>
+          <input
+            type="url"
+            name="facebook_url"
+            className="w-full px-4 py-3 border border-gray-300 rounded-xl focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all"
+            placeholder="https://facebook.com/yourprofile"
+            value={formData.facebook_url}
+            onChange={handleChange}
+          />
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default ContactTab;
