+++ Client/src/pages/UserProfile/SuccessScreen.jsx
@@ -1,0 +1,23 @@
+import React from 'react';
+
+/**
+ * SuccessScreen Component
+ * Displays success message after profile submission
+ */
+function SuccessScreen() {
+  return (
+    <div className="min-h-screen bg-gradient-to-br from-blue-50 to-indigo-100 flex items-center justify-center">
+      <div className="bg-white rounded-2xl shadow-xl p-8 text-center max-w-md">
+        <div className="w-16 h-16 bg-green-100 rounded-full flex items-center justify-center mx-auto mb-4">
+          <svg className="w-8 h-8 text-green-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M5 13l4 4L19 7" />
+          </svg>
+        </div>
+        <h2 className="text-2xl font-bold text-gray-800 mb-2">Profile Updated!</h2>
+        <p className="text-gray-600">Your profile has been successfully submitted.</p>
+      </div>
+    </div>
+  );
+}
+
+export default SuccessScreen;
